feat(cluster) new cluster events module
This module is an abstract framework that allows for broadcasting and
subscribing (pub/sub) to events across all Kong nodes connected to the
same database. It works for both PostgreSQL and Cassandra.

It consist of time series entities stored as rows inside the Kong
database. This module allows for several "channels" (although we will
most likely only use one as of now, the "invalidations" channel).

For Cassandra, each channel is a separate row in the database. The user
should have appropriate replication factor settings to ensure
availability. This makes insertion and retrieval of this time series
data particularly efficient. For Postgres, each event is its own row. We
create the necessary indexes for performance reasons. The low
cardinality of the "channels" column is helpful for such indexes.

An "event" is pure text. How said text is serialized (if complex schemas
are required) is the sole responsibility of the user.

An "event" is run once and only once per node. There is an shm tracking
the already ran event (via a v4 UUID key) for a given Nginx node.

"events" are periodically polled from a single worker via a recurring
`ngx.timer`. There is no "special worker" responsible for such polling,
to prevent a worker crash from corrupting a running node (if it stops
polling for events).

To mitigate distributed systems - multi-datacenter Cassandra cluster -
the "events" are polled with a "backwards offset" (last 60s for example)
to ensure none were missed, if they were replicated too late. This means
we fetch the same events several times, but they will only run once
anyways as previously mentioned.