refactor(*) retire Serf - for good
Remove *all* reference to Serf from our various components: datastore,
CLI (commands), core, configuration, and Admin API. We also stop
spawning Serf as a daemon from the CLI, and remove commands that don't
make sense anymore (such as checking for daemon processes running).

From now on, only our new cache invalidations events mechanism is
running.

A follow-up commit will take care of introducing a migration removing
our `nodes` table in the datastore.

From now on, the need for a CLI to start and stop Kong is drastically
reduced, almost nullified: assuming the Nginx configuration is written
and exists in the prefix, Kong can start via the Nginx binary directly,
and be the foreground process in container deployments.

This opens the way to better Nginx integration, such as described in

:fireworks: